% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheapcombfun.R
\name{CheapCombFun}
\alias{CheapCombFun}
\title{A function to calculate the cheapest drug combination given a data.frame with combination and prices}
\usage{
CheapCombFun(
  target,
  combi,
  total_max = 10,
  only_total = TRUE,
  vial_share = FALSE
)
}
\arguments{
\item{target}{The target dose}

\item{combi}{A complete dataframe with combinations and prices}

\item{total_max}{A maximum of possible combinations. Default = 10}

\item{only_total}{If \code{TRUE}, only display the total costs. If \code{FALSE}, return all information.}

\item{vial_share}{If \code{TRUE}, vial share is allowed.}
}
\description{
A function to calculate the cheapest drug combination given a data.frame with combination and prices
}
\keyword{CADTH}
