% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tx_schedule.R
\name{create_tx_schedule}
\alias{create_tx_schedule}
\title{Create a drug treatment schedule data.frame}
\usage{
create_tx_schedule(
  cyc_dur_d,
  cyc_n,
  tx1_d = 0,
  tx2_d = 0,
  tx3_d = 0,
  tx4_d = 0,
  tx5_d = 0,
  tx_names = c("foo", "bar"),
  phase_n = "Phase I"
)
}
\arguments{
\item{cyc_dur_d}{The cycle duration in days}

\item{cyc_n}{The number of cycles}

\item{tx_names}{A character vector with the respective treatment names}

\item{phase_n}{The phase if treatment changes between different phases and more than one phase is needed.}

\item{tx.._d}{Days of cycle at which drug is administered (up to 5 drugs can be added)}
}
\description{
Create a drug treatment schedule data.frame
}
\keyword{CADTH}
