% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2t.R
\name{d2m}
\alias{d2m}
\title{A set of functions to convert time}
\usage{
d2m(days)
}
\arguments{
\item{days}{The days to convert}

\item{months}{The monthts to convert}

\item{years}{The years to convert}
}
\description{
Functions to convert month to days, days, to month, years to days, days to years, years to months and months to years. Assuming that: one month contains 30.43 days on average, one year contains 365.25 days on average, and one year contains 52.17 weeks on average.
}
