% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_p_wrk_down.R
\name{f_cbs_p_work_down}
\alias{f_cbs_p_work_down}
\title{A wrapper function to download CBS data for the CPI.}
\usage{
f_cbs_p_work_down(
  save_cbs_data = TRUE,
  save_dir = "data/cbs/",
  file_name = "cbs_p_wrk.RDS"
)
}
\arguments{
\item{save_cbs_data}{Should CPI data be saved? Standard is TRUE}

\item{save_dir}{Directory where data should be saved. Standard is "data/cbs/"}

\item{file_name}{File name for the saved CPI data. Standard is "cbs_cpi.RDS}
}
\description{
A wrapper function to download CBS data for the CPI.
}
\keyword{CBS,}
\keyword{CPI}
\keyword{Generic,}
