% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_depreciation_interest.R
\name{f_depreciation_interest}
\alias{f_depreciation_interest}
\title{A function to calculate the costs of medical equipment}
\usage{
f_depreciation_interest(
  v_replace_val,
  r_salvage_val,
  n_amortisation_period = 10,
  i_interest_rt = 0.042
)
}
\arguments{
\item{v_replace_val}{V: vervangingswaarde; replacement value}

\item{r_salvage_val}{R: restwaarde; salvage value}

\item{n_amortisation_period}{N: afschrijvingstermijn,; amortization period}

\item{i_interest_rt}{i: rentepercentage; interest rate}
}
\description{
A function to calculate the costs of medical equipment based on Section 3.3 of the Dutch EE guideline; k = annual depreciation and interest expense \link{jaarlijkse afschrijvings- en rentekosten}
}
\keyword{Generic,}
\keyword{costs}
\keyword{equipment}
