% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_flexsurv_plots.R
\name{f_flexsurv_plots}
\alias{f_flexsurv_plots}
\title{Survival Analysis Plotting Function}
\arguments{
\item{fit}{A list of fitted survival models. Default is ls_obj.}

\item{show_dists}{A character vector specifying which distributions to show. Default is "all".}

\item{time_plot}{A numeric vector specifying the time points for the plot. Default is 0:6000.}

\item{conf_int_km}{A logical indicating whether to include confidence intervals for the Kaplan-Meier curve. Default is FALSE.}

\item{scale_time}{A numeric value to scale time from days to years. Default is 365.25.}
}
\value{
A ggplot2 object.
}
\description{
This function plots Kaplan-Meier and parametric survival curves.
}
\examples{
f_flexsurv_plots()
}
\author{
F. Thielen
}
