% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_word_tbl_eshpm.R
\name{f_word_tbl_eshpm}
\alias{f_word_tbl_eshpm}
\title{A wrapper function for flextable related funtctions}
\usage{
f_word_tbl_eshpm(
  df,
  cpt = "",
  tb_lbl = knitr::opts_current$get()$label,
  col_names = NULL,
  f_format_dbl = TRUE,
  f_digi = 2,
  flex_layout = "autofit"
)
}
\arguments{
\item{df}{A data.frame of tibble}

\item{cpt}{The table caption.}

\item{tb_lbl}{A table label used for bookdown documents. Default is the chunk label. If no chunk label provided add own here.}

\item{col_names}{New column names. If left empty (NULL) names will be taken from \code{df}.}

\item{f_format_dbl}{If TRUE (default) numbers in table will be formatted.}

\item{f_digi}{Number of digits when \code{f_format_dbl} is TRUE.}

\item{flex_layout}{Used for \code{flextable::set_table_properties}. 'autofit' or 'fixed' algorithm. Default to 'autofit'.}
}
\description{
A wrapper function for flextable related funtctions
}
