% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_tx_schedule.R
\name{fill_create_tx_schedule}
\alias{fill_create_tx_schedule}
\title{Join phases created with \code{tatooheen::create_tx_schedulde} or \code{tatooheen::fill_create_tx_schedulde}}
\usage{
fill_create_tx_schedule(
  regimen_name = "CarDex",
  regimen_phase = "I",
  df = df_tx_regimen,
  cyc_max = 100
)
}
\arguments{
\item{tx_schedule_ph1}{A treatment schedule for phase I}

\item{tx_schedule_ph2}{A treatment schedule for phase II}

\item{tx_schedule_ph3}{A treatment schedule for phase III}
}
\description{
Join phases created with \code{tatooheen::create_tx_schedulde} or \code{tatooheen::fill_create_tx_schedulde}
}
\keyword{CADTH}
