% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_create_tx_schedule.R
\name{fill_create_tx_schedule1}
\alias{fill_create_tx_schedule1}
\title{A wrapper function for \code{tatooheen::create_tx_schedule()} for pre-specified \code{tx_regimens.xlsx} tables}
\usage{
fill_create_tx_schedule1(
  regimen_name,
  regimen_phase,
  df = df_tx_regimen,
  cyc_max = 100
)
}
\arguments{
\item{regimen_name}{The regimen name for which to fill \code{tatooheen::create_tx_schedule()}}

\item{regimen_phase}{The regimen phase for which to fill \code{tatooheen::create_tx_schedule()}}

\item{df}{A data.frame in the form of \code{tx_regimens.xlsx}}

\item{cyc_max}{The maximum number of cycles when no cycle duration is given (i.e. until progression in some instances)}
}
\description{
A wrapper function for \code{tatooheen::create_tx_schedule()} for pre-specified \code{tx_regimens.xlsx} tables
}
