% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fun_darth.R
\name{fit_fun}
\alias{fit_fun}
\title{Fit parametric survival functions}
\usage{
fit_fun(data, formula)
}
\arguments{
\item{data}{The data set}

\item{formula}{The formula with or without co-variates (to be specified with \code{as.formula()})}
}
\description{
A function to fit several parametric survival functions. This function is adapted from \code{daarthools::fit.fun()}
}
