% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkbl.R
\name{fkbl}
\alias{fkbl}
\title{A wrapper function for kableExtra::kbl()}
\usage{
fkbl(df, digi = 2, big_mark = TRUE, ...)
}
\arguments{
\item{df}{A data.frame to display}

\item{digi}{The number of digits to show}
}
\description{
A wrapper function for kableExtra::kbl()
}
