% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkbl_style.R
\name{fkbl_style}
\alias{fkbl_style}
\title{A wrapper function for kableExtra::kable_styling() to be used after tatooheen::fkbl()}
\usage{
fkbl_style(
  fkbl,
  scale = FALSE,
  boots_opts = c("striped", "condensed", "hover"),
  latx_opts = c("striped", "HOLD_position"),
  ...
)
}
\arguments{
\item{fkbl}{A piped object from tatooheen::kbl()}

\item{scale}{If \code{TRUE}, table will be scaled down}

\item{boots_opts}{Standard bootstrap options for HTML. Can be overwritten.}

\item{latx_opts}{Standard options for PDF. Can be overwritten.}

\item{...}{Further arguments to \code{kableExtra::kable_styling()}}
}
\description{
A wrapper function for kableExtra::kable_styling() to be used after tatooheen::fkbl()
}
