% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_tx_schedule.R
\name{join_tx_schedule}
\alias{join_tx_schedule}
\title{Join phases created with \code{tatooheen::create_tx_schedulde} or \code{tatooheen::fill_create_tx_schedulde}}
\usage{
join_tx_schedule(tx_schedule_ph1, tx_schedule_ph2, tx_schedule_ph3 = NULL)
}
\arguments{
\item{tx_schedule_ph1}{A treatment schedule for phase I}

\item{tx_schedule_ph2}{A treatment schedule for phase II}

\item{tx_schedule_ph3}{A treatment schedule for phase III}
}
\description{
Join phases created with \code{tatooheen::create_tx_schedulde} or \code{tatooheen::fill_create_tx_schedulde}
}
\keyword{CADTH}
