% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hmd_lifetab.R
\name{load_hmd_lifetab}
\alias{load_hmd_lifetab}
\title{Load HMD data for background survival}
\usage{
load_hmd_lifetab(
  path_to_file = "data/hmd/",
  force_update = FALSE,
  country = "NLD",
  hmd_user = Sys.getenv("HMD_USER"),
  hmd_password = Sys.getenv("HMD_PW")
)
}
\arguments{
\item{path_to_file}{A character for path to file. Default is \code{"data/hmd/"}. File name will be given automatically.}

\item{force_update}{If \code{TRUE} update will be forced (i.e. new download).}

\item{country}{Country for which HMD data should be downloaded. Default is "NL".}

\item{hmd_user}{HMD username.}
}
\description{
A wrapper function for \code{HMDHFDplus::readHMDweb} to load HMD data (both male and female) for the first time. If a file for the year (based on \code{Sys.Date()}) already exists, the download is skipped. Updates can be forced with \code{force_update = TRUE}.
See https://towardsdatascience.com/using-dotenv-to-hide-sensitive-information-in-r-8b878fa72020 how to set and load environment data for username and password.
}
\keyword{HMD}
