% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_price.R
\name{p_price}
\alias{p_price}
\title{A function to write pretty prices in \code{bookdown} reports}
\usage{
p_price(x, currency = "EUR")
}
\arguments{
\item{x}{A number to be printed}

\item{currency}{The name of the currency}
}
\description{
A function to write pretty prices in \code{bookdown} reports
}
