% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fun_own.R
\name{SurvExtraGroupFun}
\alias{SurvExtraGroupFun}
\title{Fit parametric survival functions own}
\usage{
SurvExtraGroupFun(
  time.event,
  event,
  grouping = 1,
  data.source,
  wgts = NULL,
  max.yr.plot = 1,
  name.plot = ""
)
}
\description{
Fit parametric survival functions own
}
