% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_schedule_plot.R
\name{tx_schedule_plot}
\alias{tx_schedule_plot}
\title{A function to visualise treatment schedules created with for \code{tatooheen::create_tx_schedule()}}
\usage{
tx_schedule_plot(data, cycles)
}
\arguments{
\item{data}{The treatment schedule from \code{tatooheen::create_tx_schedule()}}

\item{cycle}{A selection of cycles to display}
}
\description{
A function to visualise treatment schedules created with for \code{tatooheen::create_tx_schedule()}
}
\keyword{CADTH}
