% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_create_tx_schedule.R, R/join_tx_schedule.R
\name{fill_create_tx_schedule}
\alias{fill_create_tx_schedule}
\title{A wrapper function for \code{tatooheen::create_tx_schedule()} for pre-specified \code{tx_regimens.xlsx} tables}
\usage{
fill_create_tx_schedule(
  regimen_name = "CarDex",
  regimen_phase = "I",
  df = df_tx_regimen,
  cyc_max = 100
)

fill_create_tx_schedule(
  regimen_name = "CarDex",
  regimen_phase = "I",
  df = df_tx_regimen,
  cyc_max = 100
)
}
\arguments{
\item{regimen_name}{The regimen name for which to fill \code{tatooheen::create_tx_schedule()}}

\item{regimen_phase}{The regimen phase for which to fill \code{tatooheen::create_tx_schedule()}}

\item{df}{A data.frame in the form of \code{tx_regimens.xlsx}}

\item{cyc_max}{The maximum number of cycles when no cycle duration is given (i.e. until progression in some instances)}

\item{tx_schedule_ph1}{A treatment schedule for phase I}

\item{tx_schedule_ph2}{A treatment schedule for phase II}

\item{tx_schedule_ph3}{A treatment schedule for phase III}
}
\description{
A wrapper function for \code{tatooheen::create_tx_schedule()} for pre-specified \code{tx_regimens.xlsx} tables

Join phases created with \code{tatooheen::create_tx_schedulde} or \code{tatooheen::fill_create_tx_schedulde}
}
\keyword{CADTH}
