% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkbl.R
\name{fkbl}
\alias{fkbl}
\title{A wrapper function for kableExtra::kbl}
\usage{
fkbl(df, digi = 2, big_mark = TRUE, ...)
}
\arguments{
\item{Test}{foo}
}
\description{
A wrapper function for kableExtra::kbl
}
\keyword{fkbl()}
\keyword{kable,}
\keyword{kableExtra}
